Add Publicly Verifiable Token with Metadata
Draft 2 https://datatracker.ietf.org/doc/html/draft-hendrickson-privacypass-public-metadata-03#name-client-to-issuer-request-2

Using Partially Blind RSA implementation.
Given the protocol is close to the normal Public issuance, the class is
abstracted, and two children of this abstract class are exported: one
with Metadata and the other without.

There are no test vectors yet for the protocol, which is to be added in
a following commit.

Add Origin class for publicly verifiable token

Add example for metadata

Add test vectors for Privacy Pass with metadata

Not yet part of the draft, but allow to confirm issues with the
implementation.

Add Partially Blind RSA as supported token in README