Update arbitrary batched tokens to match latest draft (v5)
Update consist f removing the length prefix to tokens. Only the overall
array is prefixed with a quicvarint.

This commit introduces a ne method to infer the tokenrequest serialized
length from the TokenTypeEntry. Should this be part of the
TokenTypeEntry directly, not sure. Maybe a method on it.

Tests have been updated with Go and Rust implementation, and interrop is
confirmed.

Hopefully the latest draft update before RFC.